Memory and C++ correctness: bounds on every indexed public accessor, ownership
made explicit where two owners or none shared a block, resampling and histogram
kernels that handled the degenerate sizes real data produces, the array library
told to saturate rather than wrap, and the Python side stopped writing into
constants the interpreter shares.

Verified: clean rebuild green, 15 test targets pass, and the same 15 pass under
AddressSanitizer with no defect reported.

+++ src/Core/VipCore.h
@@ -521,7 +521,7 @@ public:
 /// This function works for standard data types: numerical values, QString and QByteArray, VipNDArray, VipIntervalSampleVector, VipPointVector, etc.
 /// It is possible to define additional memory computation functions for custom types using vipRegisterMemoryFootprintFunction().
 /// If the data type is not handled, 0 is returned.
-VIP_CORE_EXPORT int vipGetMemoryFootprint(const QVariant& v);
+VIP_CORE_EXPORT qint64 vipGetMemoryFootprint(const QVariant& v);
 
 /// Register, for a given Qt meta type id, a function that computes the memory footprint of a QVariant.
 /// This function will be used in vipGetMemoryFootprint().
